+++ resources/views/components/form/bsRadio.blade.php
@@ -1,6 +1,4 @@
-<div class="form-check">
-        <label class="form-check-label">
-                {{ Form::radio($name, $value, $checked, [ 'class' => 'form-check-input' ]) }}
-                {{ $label }}
-        </label>
+<div class="custom-control custom-radio">
+        {{ Form::radio($name, $value, $checked, [ 'class' => 'custom-control-input', 'id' => form_id_string($name) ]) }}
+        <label class="custom-control-label" for="{{ form_id_string($name) }}">{{ $label }}</label>
 </div>