+++ resources/views/components/form/bsRadioList.blade.php
@@ -2,7 +2,10 @@
     <p>{{ $label }}</p>
 @endif
 @foreach($entries as $k => $v)
-@include('components.form.bsRadio', [ 'name' => $name, 'value' => $k, 'checked' => $k == $value, 'label' => $v ])
+    <div class="custom-control custom-radio">
+        {{ Form::radio($name, $k, $k == $value, [ 'class' => 'custom-control-input', 'id' => form_id_string($name, $k) ]) }}
+        <label class="custom-control-label" for="{{ form_id_string($name, $k) }}">{{ $v }}</label>
+    </div>
 @endforeach
 @if ($errors->has($name))
     <div><small class="text-danger">{{ $errors->first($name) }}</small></div>