+++ resources/views/roles/create.blade.php
@@ -27,15 +27,10 @@
                 <div class="card">
                     <div class="card-header">@lang('app.permissions')</div>
                     <div class="card-body">
-
-                        @forelse ($permissions as $k => $v)
-                            <label>
-                                {{ Form::checkbox('permissions[]', $k) }} {{ $v }}
-                            </label><br>
-                        @empty
+                        {{ Form::bsCheckboxList('permissions[]', $permissions, null) }}
+                        @empty($permissions)
                             <em>@lang('app.no_permissions')</em>
-                        @endforelse
-
+                        @endempty
                     </div>
                 </div>
             </div>