+++ resources/views/users/create.blade.php
@@ -38,13 +38,10 @@
                 <div class="card">
                     <div class="card-header">@lang('app.roles')</div>
                     <div class="card-body">
-                        @forelse ($roles as $role)
-                            <label>
-                                {{ Form::checkbox('roles[]', $role->id) }} {{ $role->name }}
-                            </label><br>
-                        @empty
+                        {{ Form::bsCheckboxList('roles[]', $roles->mapWithKeys(function($role){return [ $role->id => $role->name ];}), null) }}
+                        @empty($roles)
                             <em>@lang('app.no_roles')</em>
-                        @endforelse
+                        @endempty
                     </div>
                 </div>
             </div>